Add blazer gem
And run `rails generate blazer:install`

This resulted in errors due to the app only using railties instead of the full Rails. Removing the unused Rails portions allowed the blazer installation to proceed.

I tested being able to connect to and query a remote DB by using one of my Heroku apps. I was not able to connect to the `int` IdP DB, even with Mossadeq's help, where we tried allowing access from an IP address within the GSA network.

In `config/blazer.yml`, I added `smart_columns` for `event_type` so that we can see the names of the events when we query a user's events. We can make those more descriptive later if we need to.